+++ NEWS.md
@@ -1,6 +1,7 @@
 ## storr 1.1.3 (201?-??-??)
 
 * Only use version 2 serialisation, avoiding breakage when version 3 is released in 3.5.0 (#62, reported by Tomas Kalibera)
+* Quote sql more safely (#60, reported by @wlandau)
 
 ## storr 1.1.2 (2017-09-08)